+++ feabas_workbench/__init__.py
@@ -1,5 +1,9 @@
 """FEABAS Workbench: a desktop front end around FEABAS for serial-section EM."""
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 APP_NAME = "FEABAS Workbench"
 ORG_NAME = "FeabasWorkbench"
+WORKBENCH_REPO = "https://github.com/georgkislinger/feabas-workbench"
+FEABAS_REPO = "https://github.com/YuelongWu/feabas"
+# Wu, Y. & Lichtman, J. W. (2026). FEABAS: A Stitching and Alignment Tool for Serial EM Data. bioRxiv.
+FEABAS_PAPER = "https://doi.org/10.64898/2026.06.07.730510"